Removing duplicated files and adding auth into the main app structure
