+++ README.md
@@ -66,6 +66,71 @@ def view_categories():
                             categories=categories)
 ```
 
+# Directory Structure
+
+Below is an outline of what this project's directory structure can and will look like. Please note that working with the file system is not the focus of this project, but rather the planning of its' directory structure.
+
+Please feel free to add, edit and leave comments and/or notes in this file as you see fit. 
+
+Please do not alter the general layout of this template, as markdown can be very sensitive and small spaces, characters, etc. can throw off the whole layout.
+
+
+    __/business-analysis-project
+    |
+    |____ app
+        |___ __init__.py 
+        |___ api.py 
+        |___ authentication.py
+        |___ categories.py
+        |___ config.py
+        |___ home.py
+        |___ organisations.py
+        |____ templates
+            |____ api
+                |___ list_organisations.html
+            |____ authentication
+                |___login.html
+                |___register.html		
+                |___logout.html
+                |___reset_password.html
+                |___user_dashboard.html
+            |____ categories
+                |___ clustering.html
+            |____ home
+                |___ home.html
+            |____ organisations
+                |___ create_organisations.html
+                |___ edit_organisations.html
+                |___ list_organisations.html
+            |___ base.html
+            |___ navbar.html
+        |____ functions
+            |___ __init__.py
+            |___ company_web_scraper.py
+            |___ create_cat_from_nace.py
+            |___ text_rank.py
+        |____ models
+            |___ __init__.py
+            |___ user.py
+        |____ static
+            |___ css
+                |___ style.css
+            |___ js
+                |___ script.js
+                |___ main.js
+    |____ data
+        |___ nace_codes.json
+    |____ docs
+        |___
+    |___ .gitignore
+    |___ .gitpod.yml
+    |___ env.py.sample
+    |___ LICENSE
+    |___ Procfile
+    |___ README.md
+    |___ requirements.txt
+    |___ run.py
+
 # Challenge
 
 - The primary challenge is to query an existing company name dataset, find their website, extract the relevant data which can be queried by search.
