+++ app/templates/auth/logout.html
@@ -15,7 +15,7 @@
 	</div>
 	<div class="row d-flex text-center">
 		<div class="col-12">
-			<form method="POST" class="logout-form" action="{{ url_for('logout') }}">
+			<form method="POST" class="logout-form" action="{{ url_for('auth.logout') }}">
 				<input type="submit" class="btn btn-primary" value="Logout">
 				
 			</form>
