+++ app/templates/auth/reset_password.html
@@ -1,8 +1,8 @@
 {% extends "base.html" %}
-{% from "_formhelpers.html" import render_field %}
+{% from "auth/_formhelpers.html" import render_field %}
 
 {% block content %}
-<form name="reset-password" method="POST" class="reset-password" action="{{ url_for('reset_password') }}">
+<form name="reset-password" method="POST" class="reset-password" action="{{ url_for('auth.reset_password') }}">
 	{{ form.hidden_tag() }}
 	{{ render_field(form.username) }}
 	{{ render_field(form.password) }}
@@ -17,7 +17,7 @@
 	<div class="row d-flex text-center">
 		<div class="col-12">
 			<p>Want to login instead?</p>
-		<a class="link-text" href="{{ url_for('login') }}">Login</a>
+		<a class="link-text" href="{{ url_for('auth.login') }}">Login</a>
 	</div>
 </form>
 {% endblock %}
