+++ app/templates/base.html
@@ -19,9 +19,15 @@
 	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 	<!-- Bootstrap Icons -->
 	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css">
-	<title>Business Analysis{% block title %}{% endblock %}</title>
+    
+	<title>{% block title %}Business Analysis{% endblock %}</title>
+
 	{% block metas %}
 	{% endblock %}
+
+    {% block styles %}
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
+    {% endblock %}
 </head>
 
 <body>
@@ -35,6 +41,7 @@
         {% with messages = get_flashed_messages() %}
         {% if messages %}
         {% for message in messages %}
+        <div class="alert alert-info flash-message">{{message}}</div>
         {% endfor %}
         {% endif %}
         {% endwith %}
