--- run
@@ -1,1 +1,0 @@
-FLASK_APP=app.py FLASK_ENV=development flask run
