--- static/css/style.css
@@ -1,60 +1,0 @@
-a.link-text {
-	text-decoration: none;
-	color: rgb(41, 158, 158);
-
-}
-
-a.link-text:hover {
-	color: rgb(16, 99, 99);
-}
-
-
-/* Authentication Forms */
-
-#registerForm {
-	align-items: center;
-	padding-top: 40px;
-	padding-bottom: 40px;
-}
-
-main.form-register,
-main.form-login {
-	display: block;
-}
-
-.card-wrapper {
-	display: flex;
-	justify-content: center;
-	min-width: 100%;
-	max-width: 800px;
-	height: auto;
-	padding: 40px 40px;
-	margin: auto;
-}
-
-.card {
-	margin-top: 80px;
-	-webkit-box-shadow: 0 10px 6px -6px #777;
-	-moz-box-shadow: 0 10px 6px -6px #777;
-	box-shadow: 0 10px 6px -6px #777;
-	border: .5px solid rgb(194, 194, 194);
-}
-
-form.login-form,
-form.register-form {
-	margin: 30px 80px;
-	display: block;
-}
-
-#loginButton,
-#registerButton {
-	margin-top: 20px;
-	margin-top: 20px;
-	width: fit-content;
-}
-
-#buttonRow {
-	align-items: center;
-	display: flex;
-	justify-content: center;
-}
