--- templates/base.html
@@ -1,128 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="utf-8">
-	<meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<meta name="robots" content="index, follow">
-	<meta name="keywords" content="business, analysis, analytics, county, council, Dún, Laoghaire-Rathdown{% block extra_keywords %}{% endblock %}">
-	<!-- Facebook, LinkedIn, Google -->
-	<meta property="og:type" content="article" />
-	<meta property="og:title" content="Business Analysis{{ self.title() }}" />
-	<meta property="og:site_name" content="Business Analysis{{ self.title() }}" />
-	<!-- Twitter -->
-	<meta name="twitter:card" content="summary">
-	<meta name="twitter:title" content="Business Analysis{{ self.title() }}">
-	<!-- Bootstrap5.1 CDN -->
-	<!-- Compiled and minified CSS -->
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
-		integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous"> 
-	<!-- Bootstrap Icons -->
-	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css">
-	<!--Font Awesome-->
-	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css"
-		integrity="sha512-9usAa10IRO0HhonpyAIVpjrylPvoDwiPUiKdWk5t3PyolY1cOd4DSE0Ga+ri4AuTroPR5aQvXU9xC6qOPnzFeg=="
-		crossorigin="anonymous" referrerpolicy="no-referrer" />
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-	{% block title %}
-	<title>Business Analysis Project</title>
-	{% endblock %}
-
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-	<!-- Bootstrap Icons -->
-	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css">
-	{% block metas %}
-	{% endblock %}
-</head>
-
-{% block styles %}
-<link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
-{% endblock %}
-
-<body>
-	<!--Start Header-->
-	<header>
-		<nav class="navbar navbar-expand-lg navbar-light bg-light">
-			<div class="container-fluid">
-				<a class="navbar-brand" href="{{ url_for('index') }}">Home</a>
-				<button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-					data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
-					aria-expanded="false" aria-label="Toggle navigation">
-					<span class="navbar-toggler-icon"></span>
-				</button>
-				<div class="collapse navbar-collapse" id="navbarSupportedContent">
-					<ul class="navbar-nav me-auto mb-2 mb-lg-0">
-						<li class="nav-item">
-							<a class="nav-link active" aria-current="page" href="#">Home</a>
-						</li>
-						{% if session.user %}
-						<li class="nav-item">
-							<a class="nav-link" href="{{ url_for('logout') }}">Logout</a>
-						</li>
-						{% else %}
-						<li class="nav-item">
-							<a class="nav-link" href="{{ url_for('register') }}">Register</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="{{ url_for('login') }}">Login</a>
-						</li>
-						{% endif %}
-					</ul>
-					<form class="d-flex">
-						<input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-						<button class="btn btn-outline-success" type="submit">Search</button>
-					</form>
-				</div>
-			</div>
-		</nav>
-	</header>
-	<!--./end-header-->
-
-	<!-- Flash messages -->
-	<section>
-		{% with messages = get_flashed_messages() %}
-		{% if messages %}
-		{% for message in messages %}
-		<div class="row flashes">
-			<h4 class="text-primary text-center">{{ message }}</h4>
-		</div>
-		{% endfor %}
-		{% endif %}
-		{% endwith %}
-	</section>
-
-<body>
-	<!--Start Header-->
-	<header>
-		<!--Main Navigation Goes Here-->
-		<!--./end-main-navigation-->
-	</header>
-	<!--./end-header-->
-	<!--Start Main Content-->
-	<main class="container">
-		<!--Main Content Goes Here-->
-		{% block content %}
-		{% endblock %}
-	</main>
-	<!--./end-main-content-->
-	<!--Start Footer-->
-	<footer></footer>
-	<!--./end-footer-->
-
-	<!-- Bootstrap Bundle JavaScript -->
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
-		integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
-		crossorigin="anonymous"></script>
-	<!-- jQuery -->
-	<script src="https://code.jquery.com/jquery-3.6.0.min.js"
-		integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
-	<!-- App's custom Javascript -->
-	<script src="{{ url_for('static', filename='js/scripts.js')}}"></script>
-
-	<!-- Block to implement custom scripts for template specific -->
-	{% block javascripts %}
-	{% endblock %}
-</body>
-</html>
-
