--- templates/index.html
@@ -1,7 +1,0 @@
-{% extends "base.html" %}
-
-
-{% block content %}
-<h1>Hello World</h1>
-
-{% endblock %}
